Added random number generator
This script simulates a quantum random number generator by projecting qubits to the Hadamard basis and then measure each of it in the computational basis. This gives you a bit string of random numbers {0,1}^n. Note that the probability should converge to 0.5 for an actual quantum system. As I do not know the quality of Googles simulator used here and do not know benchmarking stats for it take with a grain of salt. You can try to obtain stats though by running the simulation several times and then see the stats in a histogram. I will add this function later but first wanted to make sure that the script runs for small numbers of qubits even on outdated hardware.